Move answers feature scenarios to separate spec file.
