+++ spec/features/questions_spec.rb
@@ -1,10 +1,10 @@
 require 'rails_helper'
 
-feature 'Question' do
+feature 'Interacting with questions' do
 
   given(:user) { create(:user) }
-  given(:user1) { create(:user) }
-  given(:question) { create(:question, user: user1) }
+  given(:another_user) { create(:user) }
+  given(:question) { create(:question, user: another_user) }
 
   scenario 'Authenticated user creates question' do
     sign_in_with(user.email, user.password)
@@ -22,51 +22,20 @@ feature 'Question' do
 
   scenario 'Unauthenticated user cannot create question' do
     visit questions_path
-
     expect(page).to_not have_link 'Ask question'
   end
 
-  scenario 'Authenticated user can leave an answer to question' do
-    sign_in_with(user.email, user.password)
-    visit question_path(question)
-
-    fill_in 'Your answer', with: 'That means you gonna be Spiderman!'
-    click_on 'Add answer'
-
-    expect(page).to have_content 'Answer was added'
-    expect(page).to have_content 'That means you gonna be Spiderman!'
-  end
-
   scenario 'User can view questions list' do
-    questions = create_list(:question, 5, user: user1)
+    questions = create_list(:question, 5, user: another_user)
     visit questions_path
 
     questions.each do |question|
       expect(page).to have_link question.title
     end
-
-  end
-
-  scenario 'User or guest can see question and appropriate answers' do
-    answers = create_list(:answer, 3, question: question, user: user1)
-    visit question_path(question)
-
-    expect(page).to have_content question.title
-    expect(page).to have_content question.body
-    answers.each do |answer|
-      expect(page).to have_content answer.body
-    end
-  end
-
-  scenario 'Unauthenticated user cannot leave answers' do
-    visit question_path(question)
-
-    expect(page).to have_content 'Sign in or sign up to leave answers'
-    expect(page).to_not have_content 'Add answer'
   end
 
   scenario "Question's author can delete his question" do
-    sign_in_with(user1.email, user1.password)
+    sign_in_with(another_user.email, another_user.password)
     visit question_path(question)
 
     expect(page).to have_content 'Delete question'
@@ -78,5 +47,4 @@ feature 'Question' do
 
     expect(page).to_not have_content 'Delete question'
   end
-
 end
